+++ src/controllers/RequestController.ts
@@ -9,6 +9,7 @@ class RequestController {
 	public async creatReq(req: Request, res: Response): Promise<Response> {
 		const requests: Requests = req.body
 
+
 		const result = await getCustomRepository(RequestRepository).clockreq(requests)
 
 		if (result) {
