+++ src/utils/db-connection.ts
@@ -14,6 +14,8 @@ class DBConnection {
 				}
 			})
 			.catch((err) => {
+
+
 				console.log('failed to create connection')
 			})
 	}
